[FE] Modal 코드 분리 (#284)
* refactor: modal 상태 코드 분리

* refactor: modal type 분리

* refactor: modal 코드 분리
